pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "RESIDENCY-DAY-1-GROUP-Project"
version = "0.1.0"
description = "A chatbot that answers immigration questions thoroughly, impatiently, and politely."
authors = [
    { name = "Group1", email = "XXXX@uc.edu" }
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "django",
    "flask>=3.0.3",
    "python-dotenv",
    "requests>=2.32.3",
    "pyyaml",
    "openai==1.68.2",
    "urllib3>=2.2.3", 
    "werkzeug>=3.0.4", 
]

[tool.setuptools]
packages = ["src"] 